+++ lib/src/core/agents/full_screen_dialog_agent/full_screen_dialog_hide_params.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_easy_dialogs/src/core/agents/dialog_agent_base.dart';
+import 'package:flutter_easy_dialogs/src/core/agents/full_screen_dialog_agent/full_screen_dialog_agent.dart';
+
+/// Hide params for [FullScreenDialogAgent]
+class FullScreenHideParams extends AgentHideParams {}