+++ lib/src/core/dialogs/dialogs.dart
@@ -1,3 +1,4 @@
+export 'easy_banner/easy_banner_export.dart';
 export 'easy_dialog_position.dart';
+export 'easy_modal_banner/easy_modal_banner_export.dart';
 export 'factory/i_easy_dialog_factory.dart';
-export 'pre_built/pre_built.dart';